+++ src/core/scene.ts
@@ -1,0 +1,11 @@
+export class Scene {
+  response: string;
+  start: number;
+  end: number;
+
+  constructor(response: string, start: number, end: number) {
+    this.response = response;
+    this.start = start;
+    this.end = end;
+  }
+}