+++ src/types/search.ts
@@ -1,6 +1,21 @@
-import { SearchTypeValues } from '@/core/search';
+import { IndexTypeValues, SearchTypeValues } from '@/core/search';
 
 export type SearchType = keyof typeof SearchTypeValues;
+export type IndexType = keyof typeof IndexTypeValues;
+
+export type SceneSearchBase = {
+  query: string;
+  resultThreshold?: number;
+  scoreThreshold?: number;
+};
+
+export type SceneVideoSearch = {
+  videoId: string;
+} & SceneSearchBase;
+
+export type SceneCollectionSearch = {
+  collectionId: string;
+} & SceneSearchBase;
 
 export type SemanticSearchBase = {
   query: string;